improve error return for duplicate parameters
The constructor croaks in absence of a request method, so we need
to supply a minimal amount of information about the request.

Really the returned request object should not be devoid of all info
about the request anyway. Then again it should not look too much like
it would under a valid request, which might lead the caller to proceed
without ever checking for errors. Yet it also should be possible to do
so to some extent without instantly running into an exception, thereby
allowing error checks to accumulate at natural locations in the caller,
instead of being strewn all over. In this case, however, what would it
even mean to pass along the parameters when they did not parse validly?

For now, these design choices remain to be resolved in the future.
And maybe this patch will be turn out to be good enough.