add get_grant method to undispatched token request
This was originally absent because the caller should not try to get
the grant for a token request prior dispatching by supported grant
type, and the absence of the method would cause a crash in that case.

However, this is in contradiction to the design choice that the caller
should not need to check for dispatch errors directly, but should be
able to just check whether attempting to get the grant succeeded. That
means calling get_grant on whatever a failed dispatch returns – which
is the original request, just with an error set on it.

Therefore the bare token request class must have such a method as well,
and it must return nothing.